+++ tests/Functional/AtriumTestKernel.php
@@ -13,6 +13,7 @@ use Atrium\Tests\Fixtures\Data\SampleData;
 use Atrium\Tests\Fixtures\Resource\ActionsTagResource;
 use Atrium\Tests\Fixtures\Resource\ConfirmTagResource;
 use Atrium\Tests\Fixtures\Resource\DehydrateTagResource;
+use Atrium\Tests\Fixtures\Resource\FilteredTagResource;
 use Atrium\Tests\Fixtures\Resource\LayoutTagResource;
 use Atrium\Tests\Fixtures\Resource\PaginatedTagResource;
 use Atrium\Tests\Fixtures\Resource\TabsTagResource;
@@ -104,6 +105,10 @@ final class AtriumTestKernel extends Kernel
             ->autowire();
 
         $services->set(PaginatedTagResource::class)
+            ->autoconfigure()
+            ->autowire();
+
+        $services->set(FilteredTagResource::class)
             ->autoconfigure()
             ->autowire();